update: better separated imt topology parameters
